refactor(Country): Change format of Country Enum
