+++ src/main/scala/com/particeep/api/models/enums/Country.scala
@@ -1,16 +1,13 @@
 package com.particeep.api.models.enums
 
-object Country {
-  sealed abstract class Country extends Product with Enum
+sealed trait Country extends Product with Enum
 
+object Country extends EnumHelper[Country] {
   case object BE extends Country
   case object DE extends Country
   case object FR extends Country
   case object IT extends Country
 
-  object Country extends EnumHelper[Country] {
-    def values: Set[Country] = Set(BE, DE, FR, IT)
-
-    implicit def stringToCountry(value: String): Country = get(value).getOrElse(FR)
-  }
+  def values: Set[Country]                             = Set(BE, DE, FR, IT)
+  implicit def stringToCountry(value: String): Country = get(value).getOrElse(FR)
 }
